refactor(domain): replace float with decimal for monetary and surface fields (will help for Big Data management later on).
- Property: InitialPrice, CurrentPrice, FinalPrice -> decimal(18,2)
- Property: Surface -> decimal(10,2), YearBuilt -> int
- Offer: OfferPrice -> decimal(18,2)
- Contact: moved ContactRole from Agent/Client subclasses (DRY)

Also added a Justification document in order to write down adjustement made over the time.

+++ YMMO_BackEnd/Domain/Entities/Offer.cs
@@ -9,7 +9,7 @@ public class Offer
     public DateTime DateModified { get; set; }
     public DateTime DatePriceUpdated { get; set; }
     
-    public float OfferPrice { get; set; }
+    public decimal OfferPrice { get; set; }
     public StatusOffer Status { get; set; }
     
     // N:1 Relationships
